wrap opcode in module
